pypackages/example_mqtt/pyproject.toml
[project]
name = "example-mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yin-Chi Chan", email = "ycc39@cam.ac.uk" }
]
requires-python = "==3.13.*"
dependencies = [
    "babel>=2.17.0",
    "paho-mqtt>=2.1.0",
    "pytz>=2025.2",
]

[project.scripts]
example-mqtt = "example_mqtt:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
